Update MainActivity.java
I added the feature to either restart the game or not by updating handleMove(). after restarting you get to choose if you want to use x or o's i added that by updating the resetBoardUI() you had. an alert dialog was added for the choice. so the choice happens at the start of a game or when resetting to a new game.